feat: routine list, exercise list, create exercise, erros and success snackbars

+++ lib/infrastructure/persistence/schemes/exercise.dart
@@ -14,5 +14,5 @@ class Exercise {
   WorkoutType type = WorkoutType.time;
 
   @Name('time')
-  int time = -1;
+  double time = -1;
 }
